Lvl tweaks et optimzations
Final hour du lvl 3 modifié, reste plus qu'à tester:
- Moins long et difficile
-Tirer avec un modifier va changer la couleur du blast pour celle du modifier
- Fix lvl 1 bug de tutorial
- Edit: Final hour lvl 2 pour que ce soit plus clair que le joueur doit faire deux murs en haut et en bas

- Lvl 3: Rework du puzzle 8. Enlevé du cheese dans certains niveaux

- "No Ammo" apparaît maintenant quand le nombre d'emergency ammo est setté à 0

+++ FONCTIONS/lvls/lvl_script.h
@@ -8,12 +8,12 @@ void (*LVL1_PUZZLES[])();// declare!
 void (*LVL2_PUZZLES[])();
 void (*LVL3_PUZZLES[])();
 
-extern short gCurrentLevel;	
-extern short gUnlockedLevels[3];
-extern short gLastLvlUnlocked;
+extern int gCurrentLevel;	
+extern int gUnlockedLevels[3];
+extern int gLastLvlUnlocked;
 
-extern short gCurrentStage;	// Le sous-niveau actuel
-extern short gCurrentPuzzle[];	// Le checkpoint que le joueur à réussie à reach durant le niveau
+extern int gCurrentStage;	// Le sous-niveau actuel
+extern int gCurrentPuzzle[];	// Le checkpoint que le joueur à réussie à reach durant le niveau
 extern bool gDayStarted;	// Si le lvl à vraiment commencé
 extern bool gSkipTutorial;		
 extern bool gRefreshStage;	// Refresh un stage
